+++ packages/react/src/internal/utils.js
@@ -17,7 +17,29 @@ export const loopDeps = (deps, res) => {
 	}
 	let result = res;
 	for (let dep of deps) {
-		result = { ...result, ...getData(dep.name, dep.type) };
+		result = { ...result, [dep.name]: getData(dep.name, dep.type) };
 	}
 	return result;
 };
+
+/**
+ * function for checking whether the mask
+ * and storage values match
+ * @param {object} params
+ * @param {object} callback
+ */
+export function checkState({ current, event, exp, expLen }, callback) {
+	let success = 0;
+	for (let key in exp) {
+		if (
+			typeof current[event.name][key] !== 'undefined' &&
+			current[event.name][key] === exp[key]
+		) {
+			success += 1;
+		}
+	}
+
+	if (success === expLen) {
+		callback();
+	}
+}