+++ packages/adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @biscuit-store/adapter
 
+## 1.1.0
+### Major Changes
+
+- State and payload include to context
+- Added includeContext function
+
 ## 0.9.99
 ### Patch Changes